Split out builder
